reuse RandBytes
