+++ contrib/cmd/runkperf/commands/data/client.go
@@ -5,7 +5,9 @@ package data
 
 import (
 	"context"
+	"crypto/rand"
 	"fmt"
+	"math/big"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -59,4 +61,32 @@ func PrepareNamespace(clientset *kubernetes.Clientset, namespace string) error {
 		return fmt.Errorf("failed to create namespace %s: %v", namespace, err)
 	}
 	return nil
+}
+
+var letterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
+
+// RandBytes generates a random byte slice of length n using alphanumeric characters.
+func RandBytes(n int) ([]byte, error) {
+	if n <= 0 {
+		return nil, fmt.Errorf("length must be positive")
+	}
+
+	b := make([]byte, n)
+	for i := range b {
+		random, err := rand.Int(rand.Reader, big.NewInt(int64(len(letterRunes))))
+		if err != nil {
+			return nil, fmt.Errorf("error generating random number: %w", err)
+		}
+		b[i] = byte(letterRunes[int(random.Int64())])
+	}
+	return b, nil
+}
+
+// RandString generates a random string of length n using alphanumeric characters.
+func RandString(n int) (string, error) {
+	b, err := RandBytes(n)
+	if err != nil {
+		return "", err
+	}
+	return string(b), nil
 }
